+++ src/@types/ReactionHandler.d.ts
@@ -1,0 +1,12 @@
+import type { MessageReaction, PartialMessageReaction, PartialUser, User } from 'discord.js';
+
+declare global {
+	interface ReactionHandler {
+		execute: (context: ReactionHandlerContext) => void | Promise<void>;
+	}
+
+	interface ReactionHandlerContext {
+		reaction: MessageReaction | PartialMessageReaction;
+		user: User | PartialUser;
+	}
+}